Bit manipulation
Explanation for solution 1:
We have to find out the number of positions at which the corresponding bits are different between two integers.
So, from the XOR of the two integers, 1's represent the position where the bits were different. So the total count of 1 in the XOR is our answer.
Let's take the example of 3 & 1.
Step 1:
3 = 0 0 1 1
1= 0 0 0 1
Xor= 0 0 1 0
So, Xor contains a 1. That means the two values differs in one position only. So, 1 will be our answer.
Step 2:
Now, we have to write code to find the count of 1 in the Xor.

int countt = 0;
while(Xor>0){
countt = countt+ (Xor & 1);
Xor = Xor>>1;
}
return countt;

Xor is now 0010(2 in decimal)
So it will enter into the loop. then,
Xor = 0 0 1 0
1= 0 0 0 1
AND= 0 0 0 0
So, countt= 0+0=0
We will now right shift the Xor by 1 position.
so, new Xor=0001
So, again loop will work. then,
Xor = 0 0 0 1
1= 0 0 0 1
AND= 0 0 0 1( 1 in decimal)
So, countt= 0+1=1;
shifting Xor to the right will give Xor= 0000
So , the loop won't work again and we will return the countt.